Refactoring
- Make a clear distinction between the different factory forms: contract factory, singleton factory, and module factory.
- Add function documentation.
- Clean up other forms of unclear naming.
