+++ hardhat.config.ts
@@ -19,7 +19,7 @@ const { network } = yargs
 dotenv.config();
 const { INFURA_KEY, MNEMONIC, ETHERSCAN_API_KEY, PK } = process.env;
 
-import "./src/factory/singleton-deployment";
+import "./src/tasks/singleton-deployment";
 
 const DEFAULT_MNEMONIC =
   "candy maple cake sugar pudding cream honey rich smooth crumble sweet treat";
