+++ src/factory/index.ts
@@ -1,3 +1,4 @@
-export * from "./factory";
+export * from "./module_deployer";
+export * from "./mastercopy_deployer";
 export * from "./types";
 export * from "./constants";
