+++ src/factory/mastercopy_deployer.ts
@@ -1,34 +1,39 @@
 import { ContractFactory } from "ethers";
 import { HardhatRuntimeEnvironment } from "hardhat/types";
-import { getSingletonFactory } from "./singleton-deployment";
+import { getSingletonFactory } from "./singleton_factory";
 
 const salt =
   "0xb0519c4c4b7945db302f69180b86f1a668153a476802c1c445fcb691ef23ef16";
 
 /**
+ * Deploy a module's mastercopy via the singleton factory.
  *
- * @param hardhat Deploy a mastercopy via the singleton factory
+ * To get the same address on any chain.
+ * @param hre hardhat runtime environment
  * @param mastercopyContractFactory
  * @param args
- * @returns
+ * @returns The address of the deployed module mastercopy
  */
 export const deployMastercopy = async (
-  hardhat: HardhatRuntimeEnvironment,
+  hre: HardhatRuntimeEnvironment,
   mastercopyContractFactory: ContractFactory,
   args: Array<any>
 ) => {
-  const deploymentTx = await mastercopyContractFactory.getDeployTransaction(
-    ...args
-  );
+  const deploymentTx = mastercopyContractFactory.getDeployTransaction(...args);
 
-  console.log("initcode ready");
-
-  const singletonFactory = await getSingletonFactory(hardhat);
+  const singletonFactory = await getSingletonFactory(hre);
 
   const targetAddress = await singletonFactory.callStatic.deploy(
     deploymentTx.data,
     salt
   );
+
+  if (targetAddress == "0x0000000000000000000000000000000000000000") {
+    throw new Error(
+      "Mastercopy already deployed to target address on this network."
+    );
+  }
+
   console.log("targetAddress", targetAddress);
 
   const deployData = await singletonFactory.deploy(deploymentTx.data, salt, {
@@ -38,10 +43,10 @@ export const deployMastercopy = async (
   const recept = await deployData.wait();
   console.log("recept", recept);
 
-  if ((await hardhat.ethers.provider.getCode(targetAddress)).length > 2) {
+  if ((await hre.ethers.provider.getCode(targetAddress)).length > 2) {
     console.log(
       "Successfully deployed ModuleProxyFactory to target address! 🎉"
     );
   }
-  return recept;
+  return targetAddress;
 };
