+++ src/factory/singleton_factory.ts
@@ -1,66 +1,18 @@
-import "hardhat-deploy";
-import "@nomiclabs/hardhat-ethers";
 import { Contract } from "ethers";
-import { task } from "hardhat/config";
 import { HardhatRuntimeEnvironment } from "hardhat/types";
-
 const singletonFactoryAbi = [
   "function deploy(bytes memory _initCode, bytes32 _salt) public returns (address payable createdContract)",
 ];
 const singletonFactoryAddress = "0xce0042b868300000d44a59004da54a005ffdcf9f";
-const factorySalt =
-  "0xb0519c4c4b7945db302f69180b86f1a668153a476802c1c445fcb691ef23ef16";
-const AddressZero = "0x0000000000000000000000000000000000000000";
 
-const deployFactory = async (_: null, hardhat: HardhatRuntimeEnvironment) => {
-  const singletonFactory = await getSingletonFactory(hardhat);
-  console.log("Singleton Factory:     ", singletonFactory.address);
-  const Factory = await hardhat.ethers.getContractFactory("ModuleProxyFactory");
-  // const singletonFactory = new hardhat.ethers.Contract(singletonFactoryAddress, singletonFactoryAbi)
-
-  const targetAddress = await singletonFactory.callStatic.deploy(
-    Factory.bytecode,
-    factorySalt
-  );
-  if (targetAddress == AddressZero) {
-    console.log(
-      "ModuleProxyFactory already deployed to target address on this network."
-    );
-    return;
-  } else {
-    console.log("Target Factory Address:", targetAddress);
-  }
-
-  const transactionResponse = await singletonFactory.deploy(
-    Factory.bytecode,
-    factorySalt
-  );
-
-  const result = await transactionResponse.wait();
-  console.log("Deploy transaction:    ", result.transactionHash);
-
-  const factory = await hardhat.ethers.getContractAt(
-    "ModuleProxyFactory",
-    targetAddress
-  );
-
-  const factoryArtifact = await hardhat.artifacts.readArtifact(
-    "ModuleProxyFactory"
-  );
-
-  if (
-    (await hardhat.ethers.provider.getCode(factory.address)) !=
-    factoryArtifact.deployedBytecode
-  ) {
-    console.log("Deployment unsuccessful: deployed bytecode does not match.");
-    return;
-  } else {
-    console.log(
-      "Successfully deployed ModuleProxyFactory to target address! 🎉"
-    );
-  }
-};
-
+/**
+ * Get the singleton factory contract (ERC-2470).
+ * If it is not deployed on the newtwork, then also deploy it.
+ *
+ * https://eips.ethereum.org/EIPS/eip-2470
+ * @param hardhat
+ * @returns Singleton Factory contract
+ */
 export const getSingletonFactory = async (
   hardhat: HardhatRuntimeEnvironment
 ): Promise<Contract> => {
@@ -101,8 +53,3 @@ export const getSingletonFactory = async (
   }
   return singletonFactory;
 };
-
-task(
-  "singleton-deployment",
-  "Deploy factory through singleton factory"
-).setAction(deployFactory);
